feat#124/티켓 재고 엔티티 변경 (#130)
* feat: 티켓 재고를 여러 레코드로 생성

- 한 레코드의 컬럼에 대해 재고를 업데이트하면 Lock이 경합 상황이 자주 발생
- 재고 수만큼의 레코드를 만들어 점유하도록 수정
- DB 접근 시 배치 처리

* feat: 티켓 재고 엔티티 수정에 따른 로직 수정

- Purchase 로직 수정
- repository query 수정
- 에러 코드 추가
- 기존 decrease를 cancel로 로직 수정
- 티켓과 매칭 되지 않을 경우의 에러코드 추가
- getPurchasePreviewInfo url mapping 변경 및 path 변수 추가

* test: 티켓 재고 로직 개선에 따른 테스트 수정

- 배치 처리 테스트 작성
- remainStock 삭제
- 컨트롤러 테스트 Json 반환 값 수정
- 구매 로직 절차 및 순서 수정

* chore: DDL 수정

- TicketStock 변경에 따른 DDL 수정 및 인덱스 추가

* refactor: TicketStock PK 컬럼 이름 변경 및 TicketStockRepository 안쓰는 쿼리 삭제

* ci: docker 파일 멀티 모듈에 맞게 수정

* refactor: 네이티브 쿼리 수정

- 네이티브 서브 쿼리를 빼고 Limit를 앞에 붙여서 최적화

* config: docker-compose.yml influxdb 추가

* refactor: 리뷰 내용 수정

- 카운트 쿼리 -> Exists 쿼리로 변경
- TicketStock 메서드 이름 변경
- 테스트 수정

* chore: influx DB 도커 네트워크 연결 추가

* chore: influx DB 볼륨 추가

* chore: docker-compose 버전 삭제

* chore: prod yml 수정

---------

+++ backend/api-server/src/main/java/com/wootecam/festivals/domain/ticket/dto/TicketResponse.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
  */
 public record TicketResponse(Long id,
                              String name, String detail,
-                             Long price, int quantity, int remainStock,
+                             Long price, int quantity, Long remainStock,
                              LocalDateTime startSaleTime, LocalDateTime endSaleTime,
                              LocalDateTime refundEndTime,
                              LocalDateTime createdAt, LocalDateTime updatedAt) {
